# src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"PageMetadata": {
				"properties": {
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"totalElements": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"pageSize",
					"totalElements",
					"totalPages",
					"currentPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"OrderDir": {
				"type": "string",
				"enum": [
					"ASC",
					"DESC"
				]
			},
...
			"ErrorMessage_404_": {
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							404
						],
						"nullable": false
					},
					"code": {
						"type": "string"
					},
					"details": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"code",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend-f1",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/free-practices/{session}/results": {
			"get": {
				"operationId": "GetSessionsResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"items": {
											"items": {
												"$ref": "#/components/schemas/TimedSessionResult"
											},
											"type": "array"
										},
										"pageData": {
											"$ref": "#/components/schemas/PageMetadata"
										}
									},
									"required": [
										"items",
										"pageData"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Return driver results of a specific practice session based on some filters",
				"tags": [
					"Free Practices"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "session",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"fp1",
								"fp2",
								"fp3",
								"fp4",
								"warm-up"
							]
						}
					},
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the chassis manufacturer",
						"in": "query",
						"name": "chassisManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the  engine manufacturer",
						"in": "query",
						"name": "engineManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to the event that has this ID",
						"in": "query",
						"name": "eventId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to this year's events",
						"in": "query",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "If specified, it will return only results relative to the pilot having this ID",
						"in": "query",
						"name": "driverId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "positionText",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"default": "eventId",
							"type": "string",
							"enum": [
								"chassisManufacturerId",
								"engineManufacturerId",
								"positionText",
								"eventId",
								"driverId",
								"positionOrder",
								"position",
								"time",
								"laps",
								"entrantName",
								"chassisName",
								"engineName",
								"driverNumber",
								"note",
								"seasonEntrantId",
								"tyreManufacturerId"
							]
						}
					}
				]
			}
		},
		"/free-practices/{session}/results/{eventId}": {
			"get": {
				"operationId": "GetSessionResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TimedSessionResult"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets the results of a certain session of free practice",
				"tags": [
					"Free Practices"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "session",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"fp1",
								"fp2",
								"fp3",
								"fp4",
								"warm-up"
							]
						}
					},
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/free-practices/{session}/results/{eventId}/{driverId}": {
			"get": {
				"operationId": "GetDriverSessionResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimedSessionResult"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets info about the result obtained by a driver in a certain session of free practice",
				"tags": [
					"Free Practices"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "session",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"fp1",
								"fp2",
								"fp3",
								"fp4",
								"warm-up"
							]
						}
					},
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "driverId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/races/results": {
			"get": {
				"operationId": "GetRacesResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"items": {
											"items": {
												"$ref": "#/components/schemas/RaceResult"
											},
											"type": "array"
										},
										"pageData": {
											"$ref": "#/components/schemas/PageMetadata"
										}
									},
									"required": [
										"items",
										"pageData"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get driver race results based on some filters",
				"tags": [
					"Races"
				],
				"security": [],
				"parameters": [
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the chassis manufacturer",
						"in": "query",
						"name": "chassisManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the  engine manufacturer",
						"in": "query",
						"name": "engineManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to the event that has this ID",
						"in": "query",
						"name": "eventId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to this year's events",
						"in": "query",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "If specified, it will return only results relative to the pilot having this ID",
						"in": "query",
						"name": "driverId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by a specific grid postion text",
						"in": "query",
						"name": "gridPos",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by a specific postion text, that can be `1`, `2`, `3`... or `DNF`, `DNS`...",
						"in": "query",
						"name": "positionText",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Look for the results where the driver achieved a position worse than or equal to this number.",
						"in": "query",
						"name": "minPos",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Look for the results where the driver achieved a position better than or equal to this number.",
						"in": "query",
						"name": "maxPos",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"default": "eventId",
							"type": "string",
							"enum": [
								"chassisManufacturerId",
								"engineManufacturerId",
								"positionText",
								"eventId",
								"driverId",
								"positionOrder",
								"position",
								"time",
								"laps",
								"entrantName",
								"chassisName",
								"engineName",
								"driverNumber",
								"note",
								"seasonEntrantId",
								"tyreManufacturerId",
								"timePenalty",
								"gap",
								"reasonRetired",
								"gridPosition",
								"gridPenalty",
								"points",
								"pointsGained",
								"pointsCountForWDC"
							]
						}
					}
				]
			}
		},
		"/races/results/{eventId}": {
			"get": {
				"operationId": "GetRaceResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RaceResult"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets info about the results of a certain race",
				"tags": [
					"Races"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/races/results/{eventId}/{driverId}": {
			"get": {
				"operationId": "GetDriverRaceResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RaceResult"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets info about the result obtained by a driver in a certain race",
				"tags": [
					"Races"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "driverId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/pre-qualifyings/results": {
			"get": {
				"operationId": "GetPreQualifyingsResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"items": {
											"items": {
												"$ref": "#/components/schemas/TimedSessionResult"
											},
											"type": "array"
										},
										"pageData": {
											"$ref": "#/components/schemas/PageMetadata"
										}
									},
									"required": [
										"items",
										"pageData"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Return driver results of a specific pre-qualifying session based on some filters",
				"tags": [
					"Pre-Qualifying"
				],
				"security": [],
				"parameters": [
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the chassis manufacturer",
						"in": "query",
						"name": "chassisManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the  engine manufacturer",
						"in": "query",
						"name": "engineManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to the event that has this ID",
						"in": "query",
						"name": "eventId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to this year's events",
						"in": "query",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "If specified, it will return only results relative to the pilot having this ID",
						"in": "query",
						"name": "driverId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "positionText",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"default": "eventId",
							"type": "string",
							"enum": [
								"chassisManufacturerId",
								"engineManufacturerId",
								"positionText",
								"eventId",
								"driverId",
								"positionOrder",
								"position",
								"time",
								"laps",
								"entrantName",
								"chassisName",
								"engineName",
								"driverNumber",
								"note",
								"seasonEntrantId",
								"tyreManufacturerId"
							]
						}
					}
				]
			}
		},
		"/pre-qualifyings/results/{eventId}": {
			"get": {
				"operationId": "GetSessionResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TimedSessionResult"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets the results of a certain pre-qualifying",
				"tags": [
					"Pre-Qualifying"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/pre-qualifyings/results/{eventId}/{driverId}": {
			"get": {
				"operationId": "GetDriverSessionResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimedSessionResult"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets info about the result obtained by a driver in a pre-qualifying session",
				"tags": [
					"Pre-Qualifying"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "driverId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/qualifyings/{session}/results": {
			"get": {
				"operationId": "GetSessionsResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"items": {
											"items": {
												"$ref": "#/components/schemas/TimedSessionResult"
											},
											"type": "array"
										},
										"pageData": {
											"$ref": "#/components/schemas/PageMetadata"
										}
									},
									"required": [
										"items",
										"pageData"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Return driver results of a specific qualifying session based on some filters",
				"tags": [
					"Qualifyings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "session",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"qualifying1",
								"qualifying2"
							]
						}
					},
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "Page to retrieve, starting at 0",
						"in": "query",
						"name": "pageNo",
						"required": false,
						"schema": {
							"default": "0",
							"format": "int32",
							"type": "integer",
							"minimum": 0
						}
					},
					{
						"description": "Size of the page to retrieve",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": "10",
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "Order of the results",
						"in": "query",
						"name": "orderDir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderDir"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the chassis manufacturer",
						"in": "query",
						"name": "chassisManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, the call will return only the results where the company that has this ID appears as the  engine manufacturer",
						"in": "query",
						"name": "engineManufacturerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to the event that has this ID",
						"in": "query",
						"name": "eventId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "If specified, it will return only results relative to this year's events",
						"in": "query",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "If specified, it will return only results relative to the pilot having this ID",
						"in": "query",
						"name": "driverId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "positionText",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"default": "eventId",
							"type": "string",
							"enum": [
								"chassisManufacturerId",
								"engineManufacturerId",
								"positionText",
								"eventId",
								"driverId",
								"positionOrder",
								"position",
								"time",
								"laps",
								"entrantName",
								"chassisName",
								"engineName",
								"driverNumber",
								"note",
								"seasonEntrantId",
								"tyreManufacturerId"
							]
						}
					}
				]
			}
		},
		"/qualifyings/{session}/results/{eventId}": {
			"get": {
				"operationId": "GetSessionResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TimedSessionResult"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets the results of a certain session of qualifying",
				"tags": [
					"Qualifyings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "session",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"qualifying1",
								"qualifying2"
							]
						}
					},
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/qualifyings/{session}/results/{eventId}/{driverId}": {
			"get": {
				"operationId": "GetDriverSessionResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TimedSessionResult"
								}
							}
						}
					},
					"404": {
						"description": "Register not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets info about the result obtained by a driver in a certain session of qualifying",
				"tags": [
					"Qualifyings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "session",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"qualifying1",
								"qualifying2"
							]
						}
					},
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "driverId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/sprint-qualifyings/results/{eventId}": {
			"get": {
				"operationId": "GetSprintQualifyingResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RaceResult"
									},
									"type": "array"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets info about the results of a certain sprint qualifying",
				"tags": [
					"Sprint Qualifyings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/sprint-qualifyings/results/{eventId}/{driverId}": {
			"get": {
				"operationId": "GetDriverRaceResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RaceResult"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage_404_"
								}
							}
						}
					}
				},
				"description": "Gets info about the result obtained by a driver in a certain sprint qualifying",
				"tags": [
					"Sprint Qualifyings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "driverId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}